fastquote/requirements.txt
fastapi
uvicorn[standard]
motor
python-dotenv
scikit-learn
pandas
numpy
joblib
PyJWT
passlib[bcrypt]
python-multipart
email-validator
